+++ documentation/src/main/wiki/WhatsNew.md
@@ -1,5 +1,3 @@
 # What's new
 
-* Fully [Event-Bus driven architecture](https://github.com/Cognifide/knotx/issues/63).
-* Knot.x is now modular (see [[Knot|Knot]]).
-* Multiple fixes and examples.
+* View Knot separated to Service Knot and Handlebars Knot